finished the login and adding the topbar and nav bar screen i have added the design on the appscreen and added the floating button second update
